refactor: domain 구조 수정
