+++ src/main/java/chess/domain/piece/limited_moving_chess_piece/BlackPawn.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.limited_moving_chess_piece;
 
 import chess.domain.Color;
 import chess.domain.Movement;
@@ -6,7 +6,6 @@ import java.util.List;
 
 public class BlackPawn extends Pawn {
 
-    // TODO: 색깔에 따라 방향 달라지고, 초기 위치인가에 따라 UP_UP 가능 / 불가능, 대각선 기물 먹기 가능
     public BlackPawn() {
         super(List.of(Movement.DOWN_DOWN, Movement.DOWN, Movement.LEFT_DOWN, Movement.LEFT_UP), Color.BLACK);
     }
