+++ src/main/java/chess/domain/piece/limited_moving_chess_piece/LimitedMovingChessPiece.java
@@ -1,8 +1,9 @@
-package chess.domain.piece;
+package chess.domain.piece.limited_moving_chess_piece;
 
 import chess.domain.Color;
 import chess.domain.Movement;
 import chess.domain.Position;
+import chess.domain.piece.ChessPiece;
 import java.util.List;
 
 public abstract class LimitedMovingChessPiece implements ChessPiece {
