+++ src/main/java/chess/domain/piece/linear_moving_chess_piece/Bishop.java
@@ -1,20 +1,19 @@
-package chess.domain.piece;
+package chess.domain.piece.linear_moving_chess_piece;
 
 import chess.domain.Color;
 import chess.domain.Movement;
 import java.util.List;
 
-public class King extends LimitedMovingChessPiece {
+public class Bishop extends LinearMovingChessPiece {
 
-    public King(Color side) {
+    public Bishop(Color side) {
         super(List.of(
-                Movement.LEFT, Movement.LEFT_DOWN, Movement.UP, Movement.RIGHT,
                 Movement.LEFT_DOWN, Movement.LEFT_UP, Movement.RIGHT_DOWN, Movement.RIGHT_UP
         ), side);
     }
 
     @Override
     public String name() {
-        return "K";
+        return "B";
     }
 }
