+++ src/main/java/chess/domain/piece/linear_moving_chess_piece/Queen.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.linear_moving_chess_piece;
 
 import chess.domain.Color;
 import chess.domain.Movement;
