+++ src/main/java/chess/domain/piece/linear_moving_chess_piece/Rook.java
@@ -1,8 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.linear_moving_chess_piece;
 
 import chess.domain.Color;
 import chess.domain.Movement;
-import chess.domain.Position;
 import java.util.List;
 
 public class Rook extends LinearMovingChessPiece {
